Add edit command functionality and updated execute command logic
There is no way for a user to edit or change parameters of any
internship added to their list. Trivially the preliminary UI for the
welcome to the app message is not very intuitive.

Let's add an Edit Command class to inherit from the overall command
class and create setter methods in each internship type to allow the
user to change parameters. Let's also reformat input and output of our
to have indents and dashed lines to look more intuitive and clearer.